@@ -26,10 +26,10 @@ class MinMaxZoomPreference {
   static const unbounded = MinMaxZoomPreference(null, null);
 
   /// The preferred minimum zoom level or null, if unbounded from below.
-  final double minZoom;
+  final double? minZoom;
 
   /// The preferred maximum zoom level or null, if unbounded from above.
-  final double maxZoom;
+  final double? maxZoom;
 
   appleMaps.MinMaxZoomPreference get appleMapsZoomPreference =>
       appleMaps.MinMaxZoomPreference(this.maxZoom, this.maxZoom);
