@@ -3,24 +3,25 @@ library platform_maps_flutter;
 import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:apple_maps_flutter/apple_maps_flutter.dart' as appleMaps;
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart' as googleMaps;
-import 'package:apple_maps_flutter/apple_maps_flutter.dart' as appleMaps;
+import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platform_interface.dart';
 
 part 'src/bitmap.dart';
 part 'src/camera.dart';
-part 'src/platform_maps.dart';
-part 'src/location.dart';
-part 'src/marker.dart';
-part 'src/polyline.dart';
-part 'src/polygon.dart';
 part 'src/cap.dart';
 part 'src/circle.dart';
-part 'src/joint_type.dart';
-part 'src/pattern_item.dart';
 part 'src/controller.dart';
+part 'src/joint_type.dart';
+part 'src/location.dart';
+part 'src/marker.dart';
+part 'src/pattern_item.dart';
+part 'src/platform_maps.dart';
+part 'src/polygon.dart';
+part 'src/polyline.dart';
 part 'src/ui.dart';
